# Simpler dependency init for ZKEdDSAEventTicket