Uses the bn128 curve exposed by @cedoor to avoid our own call to `buildBabyjub`.  This seems to work, and should avoid building at least one set of uncached objects.  It's not a complete fix, though, since we still call `buildEddsa` which makes its own call to `buildBabyjub`.  A complete solution to caching these things probably needs to involve PRs to both `ffjavascript` and `circomlibjs`.

Attn: @ichub @robknight