@ichub 2023-10-17
I would be curious to see the performance impact of this measured.