@artwyman 2023-10-19
A quick characterization of performance suggests the impact here is minor but measurable.  The actual call to ensureDepsInitialized doesn't get faster by enough to be measurable.  In my two test cases below the version with this PR is actually slightly slower, but across many runs the difference is simply smaller than the variance.  The first-time proving does get faster since the bn128 curve built by ensureDepsInitialized can now be reused by groth16prove.  Proving is faster by about 100ms out of a total 1250ms.  Second-time proving isn't any better because all the caching is in place.

We'd need to also fix the `buildEddsa` call to use a cached curve to get a bigger performance improvement for proving.  This PR should also provide an improvement in prove-after-verify since the cached `bn128` curve is shared between the two cases, but I didn't test for that.

Example measurements below.  These are from unit tests running on my M2 Macbook.

Baseline:
```
ZkEdDSAEventTicketPCD ensureDepsInitialized took 561.1592907905579ms
ZKEdDSAEventTicketPCD prove took 1257.2716250419617ms
ZKEdDSAEventTicketPCD prove took 511.6267919540405ms
ZKEdDSAEventTicketPCD prove took 641.4004998207092ms
ZKEdDSAEventTicketPCD prove took 579.051374912262ms
ZKEdDSAEventTicketPCD prove took 543.0087080001831ms
```

With this PR:
```
ZkEdDSAEventTicketPCD ensureDepsInitialized took 586.3058753013611ms
ZKEdDSAEventTicketPCD prove took 1121.7990832328796ms
ZKEdDSAEventTicketPCD prove took 519.8988747596741ms
ZKEdDSAEventTicketPCD prove took 550.2074580192566ms
ZKEdDSAEventTicketPCD prove took 611.0081248283386ms
ZKEdDSAEventTicketPCD prove took 573.7634587287903ms
```