@artwyman 2023-10-19
@ichub @cedoor I think this is probably worth merging as an incremental step in the right direction.  A 10% perf improvement to first-time-proving isn't nothing, but we should aim to address this in circomlibjs and/or ffjavascript to fully solve the problem.