@cedoor 2023-10-19
> @ichub @cedoor I think this is probably worth merging as an incremental step in the right direction. A 10% perf improvement to first-time-proving isn't nothing, but we should aim to address this in circomlibjs and/or ffjavascript to fully solve the problem.

Sure, makes sense. I'm waiting for their reply here -> https://github.com/iden3/ffjavascript/issues/129 / https://github.com/iden3/ffjavascript/pull/130.

@artwyman wonder if you've seen my review's comment.