@artwyman 2023-10-19
Switching away from the `bn128` curve entirely in favor of using `eddsa.F` is much simpler and has a similar performance improvement (as you'd expect from simply skipping a step entirely).  Here are some example results:

```
ZkEdDSAEventTicketPCD ensureDepsInitialized took 456.6867079734802ms
ZKEdDSAEventTicketPCD prove took 1095.3526248931885ms
ZKEdDSAEventTicketPCD prove took 494.65033292770386ms
ZKEdDSAEventTicketPCD prove took 488.66508293151855ms
ZKEdDSAEventTicketPCD prove took 517.5688753128052ms
ZKEdDSAEventTicketPCD prove took 498.0558748245239ms
```

Force push incoming to show what this looks like.  I'll still wait for @cedoor to say if he thinks this can be simplified further.