@artwyman 2023-10-19
I suggest I go ahead and merge this PR since it's strictly simplifying the dependencies, then we follow up with more discussion on the review thread as a possible future improvement.  @cedoor @ichub do you agree?  If so, can I get an accept?